sql/catalog/resolver: fold sql/schema into this package
It's all about resolution.

Release note: None
